+++ bin/micrograd_demo.ml
@@ -1,6 +1,7 @@
 open Base
 open Ocannl
 module Tn = Arrayjit.Tnode
+module Asgns = Arrayjit.Assignments
 module IDX = Train.IDX
 module TDSL = Operation.TDSL
 module NTDSL = Operation.NTDSL
@@ -81,7 +82,9 @@ let experiment seed ~no_batch_shape_inference ~use_builtin_weight_decay () =
   let module Backend = (val Arrayjit.Backends.fresh_backend ()) in
   let device = Backend.(new_virtual_device @@ get_device ~ordinal:0) in
   let ctx = Backend.init device in
-  let routine = Train.to_routine (module Backend) ctx bindings (Seq (update.fwd_bprop, sgd)) in
+  let routine =
+    Train.to_routine (module Backend) ctx bindings (Asgns.sequence [ update.fwd_bprop; sgd ])
+  in
   Train.all_host_to_device (module Backend) routine.context scalar_loss;
   Train.all_host_to_device (module Backend) routine.context learning_rate;
   (* Stdio.print_endline "\n******** scalar_loss **********"; Tensor.print_tree ~with_id:true
@@ -122,8 +125,12 @@ let experiment seed ~no_batch_shape_inference ~use_builtin_weight_decay () =
   Train.set_on_host Changed_on_devices mlp_result.value;
   (* By using jitted.context here, we don't need to copy the parameters back to the host. *)
   let result_routine =
-    Train.to_routine (module Backend) routine.context IDX.empty
-    @@ Block_comment ("moons infer", mlp_result.forward)
+    Train.to_routine
+      (module Backend)
+      routine.context IDX.empty
+      [%cd
+        ~~("moons" "infer";
+           mlp_result.forward)]
   in
   Stdio.print_endline "\n******** mlp_result **********";
   Tensor.print_tree ~with_id:true ~with_grad:false ~depth:9 mlp_result;