+++ lib/operation.ml
@@ -150,7 +150,8 @@ let rec pointpow ?(label : string list = []) ~grad_spec p t1 : Tensor.t =
   let p_t = NTDSL.number p in
   let%cd op_asn ~v ~t1 ~t2 ~projections = v =: v1 ** v2 ~projections in
   let%cd grad_asn =
-    if Tensor.is_prohibit_grad grad_spec then fun ~v:_ ~g:_ ~t1:_ ~t2:_ ~projections:_ -> Asgns.Noop
+    if Tensor.is_prohibit_grad grad_spec then fun ~v:_ ~g:_ ~t1:_ ~t2:_ ~projections:_ ->
+      Asgns.empty_comp
     else if Float.equal p 2.0 then fun ~v:_ ~g ~t1 ~t2:_ ~projections -> g1 =+ p_t *. t1 * g
     else if Float.equal p 1.0 then fun ~v:_ ~g ~t1 ~t2:_ ~projections -> g1 =+ g
     else fun ~v:_ ~g ~t1 ~t2:_ ~projections -> g1 =+ p_t *. (t1 **. (p -. 1.)) * g
@@ -215,7 +216,7 @@ let range_of_shape ?(label = []) ?(grad_spec = Tensor.Prohibit_grad) ?batch_dims
 (** A [stop_gradient] is an identity in the forward pass and a no-op in the backprop pass. *)
 let stop_gradient ?(label = []) =
   let module NTDSL = Initial_NTDSL in
-  let grad_asn ~v:_ ~g:_ ~t1:_ ~projections:_ = Asgns.Noop in
+  let grad_asn ~v:_ ~g:_ ~t1:_ ~projections:_ = Asgns.empty_comp in
   let%cd op_asn ~v ~t1 ~projections = v =: v1 in
   Tensor.unop ~label:("stop_grad" :: label) ~transpose_op:Pointwise_un ~op_asn ~grad_asn
     ~grad_spec:Prohibit_grad
@@ -223,12 +224,13 @@ let stop_gradient ?(label = []) =
 let slice ?(label = []) ~grad_spec (batch_idx : Idx.static_symbol) t1 : Tensor.t =
   let module NTDSL = Initial_NTDSL in
   let op_asn ~v ~t1 ~projections =
-    Asgns.Fetch
-      {
-        array = v;
-        fetch_op = Slice { batch_idx; sliced = t1.Tensor.value };
-        dims = lazy (Lazy.force projections).Idx.lhs_dims;
-      }
+    Asgns.to_comp
+    @@ Fetch
+         {
+           array = v;
+           fetch_op = Slice { batch_idx; sliced = t1.Tensor.value };
+           dims = lazy (Lazy.force projections).Idx.lhs_dims;
+         }
   in
   let%cd grad_asn ~v:_ ~g ~t1 ~projections = g1 =+ g in
   Tensor.unop ~label:("@|" :: label) ~transpose_op:(Batch_slice batch_idx) ~op_asn ~grad_asn
@@ -237,14 +239,15 @@ let slice ?(label = []) ~grad_spec (batch_idx : Idx.static_symbol) t1 : Tensor.t
 let embed_symbol ?(label = []) static_sym : Tensor.t =
   let module NTDSL = Initial_NTDSL in
   let op_asn ~v ~projections =
-    Asgns.Fetch
-      {
-        array = v;
-        fetch_op = Embed_symbol static_sym;
-        dims = lazy (Lazy.force projections).Idx.lhs_dims;
-      }
+    Asgns.to_comp
+    @@ Fetch
+         {
+           array = v;
+           fetch_op = Embed_symbol static_sym;
+           dims = lazy (Lazy.force projections).Idx.lhs_dims;
+         }
   in
-  let grad_asn ~v:_ ~g:_ ~projections:_ = Asgns.Noop in
+  let grad_asn ~v:_ ~g:_ ~projections:_ = Asgns.empty_comp in
   Tensor.op ~label:("!@" :: label) ~op_asn ~grad_asn ~grad_spec:Prohibit_grad
     (Shape.make ~batch_dims:[] ~input_dims:[] ~output_dims:[ 1 ] ())
     []
@@ -316,7 +319,7 @@ module NTDSL = struct
   let counter ?(label = []) =
     let module NTDSL = Initial_NTDSL in
     let%cd op_asn ~v ~t1 ~projections = v =+ t1 ~projections in
-    let grad_asn ~v:_ ~g:_ ~t1:_ ~projections:_ = Asgns.Noop in
+    let grad_asn ~v:_ ~g:_ ~t1:_ ~projections:_ = Asgns.empty_comp in
     Tensor.unop ~label:("counter" :: label) ~transpose_op:Pointwise_un ~op_asn ~grad_asn
       ~grad_spec:Prohibit_grad
 end